Set warning level to -Wall and fix all problems relating to this.
Most important, all coordinates are unsigned integers now.

+++ src/interaction.cc
@@ -2,7 +2,7 @@
 #include "font.h"
 
 void
-Interaction::puttext(int x, int y, const char* s)
+Interaction::puttext(unsigned int x, unsigned int y, const char* s)
 {
 	for (; *s; s++) {
 		struct CHARACTER* c = &theFont[(unsigned char)*s];
@@ -18,24 +18,24 @@ Interaction::puttext(int x, int y, const char* s)
 }
 
 void
-Interaction::gettextsize(const char* s, int* h, int* w)
+Interaction::gettextsize(const char* s, unsigned int* h, unsigned int* w)
 {
 }
 
 void
-Interaction::clear(int x, int y, int h, int w)
+Interaction::clear(unsigned int x, unsigned int y, unsigned int h, unsigned int w)
 {
-	for (int j = 0; j < h; j++)
-		for (int i = 0; i < w; i++)
+	for (unsigned int j = 0; j < h; j++)
+		for (unsigned int i = 0; i < w; i++)
 			putpixel(i + x, j + y, 0);
 }
 
 int
-Interaction::getCoordinates(int* x, int* y)
+Interaction::getCoordinates(unsigned int* x, unsigned int* y)
 {
 	if (coordX == -1 && coordY == -1)
 		return 0;
-	*x = coordX; *y = coordY;
+	*x = (unsigned int)coordX; *y = (unsigned int)coordY;
 	 coordX = -1; coordY = -1;
 	return 1;
 }
